requirements.txt
jupyter
matplotlib
pandas
mxnet
gluonts
pytest
jupytext
pytest-env
python-dateutil==2.8.0
pystan
fbprophet
plotly
tqdm
seaborn
statsmodels